Module kolor::details::color::color_spaces

Constants

ACES2065_1

ACES2065-1 is a linear encoding in AP0 primaries with a D60 whitepoint.

ACES_CG

ACEScg is a linear encoding in AP1 primaries with a D60 whitepoint.

ADOBE_1998

Adobe RGB (1998) is a linear encoding in Adobe 1998 primaries with a D65 white point

ADOBE_WIDE

Adobe Wide Gamut RGB is a linear encoding in Adobe Wide primaries with a D50 white point

ALL_COLOR_SPACES

Array containing all built-in color spaces.

APPLE

Apple RGB is a linear encoding in Apple primaries with a D65 white point

BT_709

BT.709 is a linear encoding in BT.709 primaries with a D65 whitepoint.. It’s equivalent to Linear sRGB

BT_2020

BT.2020 is a linear encoding in BT.2020 primaries with a D65 white point BT.2100 has the same linear color space as BT.2020.

CIE_RGB

CIE RGB is the original RGB space, defined in CIE RGB primaries with white point E.

CIE_XYZ

CIE XYZ reference color space. Uses CIE XYZ primaries with white point D65.

DISPLAY_P3

Display P3 by Apple is a linear encoding in P3 primaries with a D65 white point

ENCODED_BT_709

Encoded BT.709 is BT.709 with the BT.709 OETF applied.

ENCODED_BT_2020

Encoded BT.2020 is BT.2020 with the BT.2020 OETF applied.

ENCODED_BT_2100_PQ

Encoded BT.2100 PQ is BT.2020 (equivalent to the linear BT.2100 space) with the Perceptual Quantizer inverse EOTF applied.

ENCODED_DISPLAY_P3

Encoded Display P3 is Display P3 with the sRGB OETF applied.

ENCODED_SRGB

Encoded sRGB is Linear sRGB with the sRGB OETF applied (also called “gamma-compressed”).

ICtCp_HLG

ICtCp_HLG is a non-linear encoding in BT.2020 primaries, with a D65 whitepoint, using the HLG transfer function

ICtCp_PQ

ICtCp_PQ is a non-linear encoding in BT.2020 primaries, with a D65 whitepoint, using the PQ transfer function

LINEAR_SRGB

Linear sRGB is a linear encoding in BT.709 primaries with a D65 whitepoint. Linear sRGB is equivalent to BT_709.

OKLAB

Oklab is a non-linear, perceptual encoding in XYZ, with a D65 whitepoint.

OKLCH

Oklch is a non-linear, perceptual encoding in XYZ, with a D65 whitepoint. It is a variant of Oklab with LCh coordinates intead of Lab.

P3_D60

P3-D60 (ACES Cinema) is a linear encoding in P3 primaries with a D60 white point

P3_THEATER

P3-DCI (Theater) is a linear encoding in P3 primaries with a P3-DCI white point

PRO_PHOTO

Pro Photo RGB is a linear encoding in Pro Photo primaries with a D50 white point