Module kolor::details::transform

Re-exports

pub use hlg::*;

pub use pq::*;

pub use ICtCp::*;

pub use hsx::*;

Modules

ICtCp

BT.2100 ICtCp

hlg

ARIB STD-B67 or “Hybrid Log-Gamma” used in BT.2100

hsx

transforms for Hue/Saturation/X color models, like HSL, HSI, HSV

pq

SMPTE ST 2084:2014 EOTF aka “Perceptual Quantizer” used in BT.2100

Structs

ColorTransform

ColorTransform represents a reference to a function that can apply a TransformFn or its inverse.

Functions

CIELAB_to_CIELCh

CIELAB_to_XYZ

CIELCh_to_CIELAB

CIELCh_to_XYZ

CIE_1960_UCS_to_XYZ

CIE_1960_UCS_to_uvV

CIE_1960_UCS_uvV_to_XYZ

CIE_1960_uvV_to_UCS

CIE_1960_uvV_to_xyV

CIE_1960_xyV_to_uvV

CIE_1964_UVW_to_XYZ

CIE_1976_Luv_to_XYZ

Oklab_to_XYZ

Oklch_to_XYZ

XYZ_to_CIELAB

XYZ_to_CIELCh

XYZ_to_CIE_1960_UCS

XYZ_to_CIE_1960_UCS_uvV

XYZ_to_CIE_1964_UVW

XYZ_to_CIE_1976_Luv

XYZ_to_Oklab

XYZ_to_Oklch

XYZ_to_xyY

bt601_oetf

bt601_oetf_inverse

sRGB_eotf

Applies the sRGB EOTF (electro-optical transfer function), which is also the direct inverse of the sRGB OETF.

sRGB_oetf

Applies the sRGB OETF (opto-eletronic transfer function), sometimes called “gamma compensation”

xyY_to_XYZ